refactor : 배팅 금액 관련 기능 수정
1. Validator 클래스에 배팅 금액의 자료형 int -> double로 수정.
2. Validator 클래스에 배팅 금액 짝수인지 확인하는 예외 처리 메소드 제거(double형이라 필요 없음).
3. Input 클래스에 배팅 금액 입력 메소드의 입력 자료형 int -> double로 수정.
4. Output 클래스에 수정 사항으로 인한 불필요한 출력 메소드 제거.
5. README.md에 변경 사항 수정(취소선 표시).

+++ src/main/java/domain/ui/Input.java
@@ -1,5 +1,5 @@
 /*
- * @(#)Input.java       0.2 2019.12.15
+ * @(#)Input.java       0.3 2019.12.15
  *
  * Copyright (c) 2019 lxxjn0
  */
@@ -12,7 +12,7 @@ import java.util.Scanner;
  * 입력과 관련된 기능을 담당할 객체
  *
  * @author JUNYOUNG LEE (lxxjn0)
- * @version 0.2 2019.12.15
+ * @version 0.3 2019.12.15
  */
 public class Input {
     /**
@@ -31,13 +31,13 @@ public class Input {
     }
 
     /**
-     * 사용자로부터 해당 Player의 베팅 금액을 입력받는 메소드.
+     * 사용자로부터 해당 Player의 배팅 금액을 입력받는 메소드.
      *
      * @return 해당 Player의 배팅 금액.
      */
-    public int receivePlayerBettingMoneyInput() {
+    public double receivePlayerBettingMoneyInput() {
         scanner = new Scanner(System.in);
-        return scanner.nextInt();
+        return scanner.nextDouble();
     }
 
     /**
